move PlainDoubleAdapter to StringUtils and optimized it in SearchIndexTool
