+++ common/src/main/java/org/opensearch/ml/common/utils/StringUtils.java
@@ -8,6 +8,8 @@ package org.opensearch.ml.common.utils;
 import static org.apache.commons.text.StringEscapeUtils.escapeJson;
 import static org.opensearch.action.ValidateActions.addValidationError;
 
+import java.io.IOException;
+import java.math.BigDecimal;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
 import java.security.AccessController;
@@ -37,12 +39,10 @@ import org.opensearch.action.ActionRequestValidationException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.gson.Gson;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
-import com.google.gson.JsonSyntaxException;
+import com.google.gson.*;
 import com.google.gson.reflect.TypeToken;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
 import com.jayway.jsonpath.JsonPath;
 import com.jayway.jsonpath.PathNotFoundException;
 import com.networknt.schema.JsonSchema;
@@ -73,9 +73,13 @@ public class StringUtils {
 
     public static final Gson gson;
 
+    public static final Gson plainDoubleGson;
+
     static {
         gson = new Gson();
+        plainDoubleGson = new GsonBuilder().registerTypeAdapter(Double.class, new PlainDoubleAdapter()).create();
     }
+
     public static final String TO_STRING_FUNCTION_NAME = ".toString()";
 
     private static final ObjectMapper MAPPER = new ObjectMapper();
@@ -595,6 +599,36 @@ public class StringUtils {
         } catch (JsonSyntaxException e) {
             log.error("Failed to parse JSON array string: {}", jsonArrayString, e);
             return Collections.emptyList();
+        }
+    }
+
+    /**
+     * Custom Gson adapter for Double type.
+     * Serializes numbers without scientific notation.
+     * Writes null for null, NaN, and Infinity values.
+     * Deserializes JSON numbers back to Double.
+     */
+    private static class PlainDoubleAdapter extends TypeAdapter<Double> {
+        @Override
+        public void write(JsonWriter out, Double value) throws IOException {
+            if (value == null) {
+                out.nullValue();
+                return;
+            }
+
+            if (value.isNaN() || value.isInfinite()) {
+                out.nullValue();
+                return;
+            }
+
+            BigDecimal bd = BigDecimal.valueOf(value).stripTrailingZeros();
+
+            out.jsonValue(bd.toPlainString());
+        }
+
+        @Override
+        public Double read(JsonReader in) throws IOException {
+            return in.nextDouble();
         }
     }
 }
